+++ tests/test_tokenization.py
@@ -1,8 +1,4 @@
-import argparse
 import logging
-import os
-import sys
-import tempfile
 import unittest
 
 from opusfilter import tokenization